SDC_HW2_Kushev/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
asttokens==2.1.0
async-timeout==4.0.2
atomicwrites==1.4.0
backports.weakref==1.0.post1
billiard==3.6.4.0
bkcharts==0.2
blinker==1.5
brotlipy==0.7.0
cachetools==5.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
commonmark==0.9.1
comtypes==1.1.10
constantly==15.1.0
croniter==1.3.5
cssselect==1.1.0
cytoolz==0.11.0
dataprep==0.4.5
et-xmlfile==1.1.0
executing==0.8.3
fastapi==0.87.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.25.0
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
gevent==21.12.0
gitdb==4.0.10
GitPython==3.1.29
h11==0.14.0
htmlmin==0.1.12
ImageHash==4.3.1
inflection==0.5.1
ipykernel==6.6.0
ipython==7.30.1
Jinja2==3.0.3
jsonpath-ng==1.5.3
kombu==5.2.4
llvmlite==0.38.0
Markdown==3.3.6
MarkupSafe==2.1.1
mccabe==0.6.1
Metaphone==0.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.2
multimethod==1.9
munkres==1.1.4
mypy-extensions==0.4.3
nbconvert==7.2.5
openpyxl==3.0.10
pandas-profiling==3.5.0
pandoc==2.3
patsy==0.5.2
pep8==1.7.1
phik==0.12.2
Pillow==9.2.0
plumbum==1.8.0
ply==3.11
protobuf==3.20.3
psutil==5.8.0
pure-eval==0.2.2
pyarrow==10.0.1
pyasn1-modules==0.2.8
pycosat==0.6.3
pycurl==7.45.1
pydantic==1.10.2
pydeck==0.8.0
PyDispatcher==2.0.5
pydot==1.4.2
pyee==8.2.2
pyls-spyder==0.4.0
Pympler==1.0.1
pyppeteer==1.0.2
pytest==7.1.2
python-crfsuite==0.9.8
python-dotenv==0.19.2
python-lsp-jsonrpc==1.0.0
python-stdnum==1.18
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
queuelib==1.5.0
rapidfuzz==2.13.2
regex==2021.11.10
rich==12.6.0
scipy==1.9.1
semver==2.13.0
shortuuid==1.0.11
smmap==5.0.0
sqlparse==0.4.3
starlette==0.21.0
statsmodels==0.13.2
streamlit==1.15.1
style==1.1.0
tables==3.6.1
tangled-up-in-unicode==0.2.0
tinycss2==1.2.1
typeguard==2.13.3
tzdata==2022.6
tzlocal==4.2
update==0.0.1
uvicorn==0.20.0
validators==0.20.0
varname==0.8.3
vega-datasets==0.9.0
vine==5.0.0
visions==0.7.5
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.2
whitenoise==5.3.0
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.8.2.2
yarl==1.8.1
zict==2.1.0
zope.event==4.5.0
